promejut rab

+++ Bio/main.py
@@ -14,14 +14,14 @@ def PrintMatrix(matrix):
 def ImportToX(matrix):
     for i in range(1, len(matrix)):
         for j in range(len(matrix[i]) - 1):
-            X[i - 1][j] = int(matrix[i][j])
+            X[i - 1][j] = float(matrix[i][j])
 
 
 # Saving to vector Y in list for Analysis, reading in csv
 def ImportToY(matrix):
     for i in range(1, len(matrix)):
         for j in range(len(matrix[i]) - 1, len(matrix[i])):
-            Y[i - 1] = int(matrix[i][j])
+            Y[i - 1] = float(matrix[i][j])
 
 
 # Open csv and save
@@ -29,7 +29,7 @@ with open('File/data2.csv', 'r') as csv_file:
     csv_reader = csv.reader(csv_file, delimiter=';')
     dataForAnalys = list(csv_reader)  # рабочий и самый подходящий вариант для дальнейшего анализа
 
-    # Defined X and Y for next PLS methosS
+    # Defined X and Y for next PLS methos
     X = [0] * (len(dataForAnalys) - 1)
     for i in range(len(dataForAnalys) - 1):
         X[i] = [0] * (len(dataForAnalys[i]) - 1)
